Moved classes to reduce number of Java packages
