+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/client/DefaultMalwareScanClient.java
@@ -19,7 +19,6 @@ import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sap.cds.feature.attachments.service.malware.client.httpclient.HttpClientProviderFactory;
 import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanResult;
 import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanResultStatus;
 import com.sap.cds.services.ServiceException;
