+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/client/HttpClientProviderFactory.java
@@ -1,7 +1,7 @@
 /**************************************************************************
  * (C) 2019-2025 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.service.malware.client.httpclient;
+package com.sap.cds.feature.attachments.service.malware.client;
 
 import org.apache.http.client.HttpClient;
