+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/client/MalwareScanClientProviderFactory.java
@@ -1,7 +1,7 @@
 /**************************************************************************
  * (C) 2019-2025 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.service.malware.client.httpclient;
+package com.sap.cds.feature.attachments.service.malware.client;
 
 import java.net.URI;
 import java.util.Map;
@@ -31,13 +31,12 @@ public final class MalwareScanClientProviderFactory implements HttpClientProvide
 	 */
 	public MalwareScanClientProviderFactory(ServiceBinding binding, ConnectionPool configuration) {
 		Map<String, Object> credentials = binding.getCredentials();
-		String url = (String) credentials.get("url");
-
-		URI serviceUrl = URI.create(url + "/scan").normalize();
 
 		BasicCredentials basic = new BasicCredentials((String) credentials.get("username"),
 				(String) credentials.get("password"));
 
+		String url = (String) credentials.get("url");
+		URI serviceUrl = URI.create(url + "/scan").normalize();
 		DefaultHttpDestination destination = DefaultHttpDestination.builder(serviceUrl)
 				.name(DefaultAttachmentMalwareScanner.MALWARE_SCAN_SERVICE_LABEL).basicCredentials(basic).build();
