+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/service/malware/client/DefaultMalwareScanClientTest.java
@@ -29,7 +29,6 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 
-import com.sap.cds.feature.attachments.service.malware.client.httpclient.HttpClientProviderFactory;
 import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanResultStatus;
 import com.sap.cds.services.ServiceException;
