+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/service/malware/client/MalwareScanClientProviderFactoryTest.java
@@ -1,4 +1,4 @@
-package com.sap.cds.feature.attachments.service.malware.client.httpclient;
+package com.sap.cds.feature.attachments.service.malware.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
